src/AI_codemod_language/manipulators/TypeScriptFileManipulator.ts
// src/manipulators/TypeScriptFileManipulator.ts

import { Project } from "ts-morph";
import { FileManipulator } from "./FileManipulator";

export class TypeScriptFileManipulator extends FileManipulator {
    executeQuery(query: string) {
        throw new Error("Method not implemented.");
    }
    private project: Project;

    constructor(filePath: string) {
        super(filePath);
        this.project = new Project();
        this.project.addSourceFileAtPath(filePath);
    }

    queryFunction(functionName: string): string {
        const file = this.project.getSourceFile(this.filePath);
        const func = file?.getFunction(functionName);
        return func?.getFullText() || "Function not found";
    }

    queryClass(className: string): string {
        // Similar implementation for querying a class
    }

    queryProperty(propertyName: string): string {
        // Similar implementation for querying a property
    }

    queryMethod(methodName: string): string {
        // Similar implementation for querying a method
    }

    // Implement other methods as necessary
}
